更新了关注功能，可以对用户进行关注与取消

+++ src/main/java/com/youren/bbs/service/PostService.java
@@ -7,13 +7,13 @@ import java.util.List;
 public interface PostService {
     int create(Post post);
     int update(Post post);
-    int updateState(int state,long id);
-    Post findById(long id);
-    List<Post> findByUserId(int userId);
+    int updateState(int state,Long id);
+    Post findById(Long id);
+    List<Post> findByUserId(Long userId);
     List<Post> findAll();
-    int updatebrowse(long postId,Integer browse);
-    int updatedifference(long postId,Integer difference);
-    int updateawesome(long postId,Integer awesome);
-    int updatereplynumber(long postId,Integer replynumber);
-    int updatecollectCount(long postId,Integer collectCount);
+    int updatebrowse(Long postId,Integer browse);
+    int updatedifference(Long postId,Integer difference);
+    int updateawesome(Long postId,Integer awesome);
+    int updatereplynumber(Long postId,Integer replynumber);
+    int updatecollectCount(Long postId,Integer collectCount);
 }
